Make EmbraceSpanImpl private to the factory
